move local-state into its own module
